Split array_view out into the util namespace.
